requirements.txt
requests
beautifulsoup4
python-dotenv
zhipuai
sniffio
